feat(domain) : Lotto 중복 숫자 검증 로직 추가
