+++ src/main/java/lotto/domain/Lotto.java
@@ -1,0 +1,32 @@
+package lotto.domain;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+public class Lotto {
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        validateSameNumbers(numbers);
+        this.numbers = numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        if (numbers.size() != 6) {
+            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
+        }
+    }
+
+    private void validateSameNumbers(List<Integer> numbers) {
+        Set<Integer> numbersSet = new HashSet<>(numbers);
+        if (numbersSet.size () != numbers.size()) {
+            throw new IllegalArgumentException("[ERROR] 로또 번호는 중복되지 않아야 합니다.");
+        }
+    }
+
+    public List<Integer> getNumbers() {
+        return List.copyOf(numbers);
+    }
+}
